# glimpse-api/openapi.yaml
openapi: 3.0.3
info:
  title: glimpse-api
  description: An API to manage transcriptions and events.
  version: 1.0.0
paths:
  /transcriptions:
    get:
      summary: List all transcriptions
      description: Retrieve a list of all transcriptions.
      responses:
        '200':
          description: A list of transcriptions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transcription'
    post:
      summary: Add a new transcription
      description: Create a new transcription entry.
      requestBody:
        description: Transcription object that needs to be added.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transcription'
      responses:
        '201':
          description: Transcription created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transcription'
        '400':
          description: Invalid input provided.

  /events:
    get:
      summary: List all events
      description: Retrieve a list of all events.
      responses:
        '200':
          description: A list of events.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
    post:
      summary: Add a new event
      description: Create a new event entry.
      requestBody:
        description: Event object that needs to be added.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
      responses:
        '201':
          description: Event created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Invalid input provided.

components:
  schemas:

    Location:
      type: object
      properties:
        text:
          type: string
          description: The written address that corresponds to the coordinates.
        latitude:
          type: number
          format: float
          description: Latitude of the location.
        longitude:
          type: number
          format: float
          description: Longitude of the location.

    Transcription:
      type: object
      required:
        - content
        - audio
        - timestamp
      properties:
        id:
          type: integer
          format: uint32
          description: The unique identifier for the Transcription
        content:
          type: string
          description: The content of the transcription.
        audio:
          type: string
          description: The path to the original recording.
        timestamp:
          type: string
          format: date-time
          description: The time the transcription was created.

    Event:
      type: object
      required:
        - description
        - timestamp
      properties:
        id:
          type: integer
          format: uint32
          description: The unique identifier for the event
        location:
          $ref: '#/components/schemas/Location'
          description: Location of the event.
        description:
          type: string
          description: A description of the event.
        timestamp:
          type: string
          format: date-time
          description: The time the event occurred.
        transcriptions:
          type: array 
          items:
            $ref: '#/components/schemas/Transcription'
            x-go-custom-tag: gorm:"column:isactive;type:bool;default:'true'"